+++ src/Page/LogIn/LoginBarrel.ts
@@ -1,11 +1,13 @@
 import Banner from '../../Component/SignUp/Banner';
 import { GroupHeader } from '../../Component/Mission/GroupHeader';
-import { ReactComponent as ArrowRight } from '../../image/Icon/Arrow/Arrow_icon_Right.svg';
 import { LoginInput } from '../../Component/Login/LoginInput';
+import LoginForm from '../../Component/Login/LoginForm';
 
 import { SignClearBtnS } from '../../StyleComp/SignBtnS';
+
+import { ReactComponent as ArrowRight } from '../../image/Icon/Arrow/Arrow_icon_Right.svg';
 
 export * from '../../StyleComp/LoginInputS';
 export * from '../../API/login';
 
-export { Banner, GroupHeader, SignClearBtnS, ArrowRight, LoginInput };
+export { Banner, GroupHeader, SignClearBtnS, ArrowRight, LoginInput, LoginForm };
